+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 with an alias in the data storage.
 - Added annotation `CompoundKey`. It could be used to define a compound key for databases 
 that only allow one partition key
+- Added sheet name into arguments of ExcelConnector
 
 ## 0.2.5 (2019-06-12)
 - Added DynamoDB V2 repository